Stop configuring the max width and height in YogaUIView (#2757)
* Stop configuring the max width and height in YogaUIView

These attributes only cause us pain.

In particular they aren't necessary to pass any of
our layout tests, and by removing them I can fix a
layout bug.

Note that on Android we do not use these attributes,
so this brings the two platforms closer together.

Closes: https://github.com/cashapp/redwood/issues/2753

* Start using systemLayoutSizeFittingSize

This way we can tell the view whether a measurement is
exactly (which it should honor unconditionally), or not
(which it might honor if its constraint is Fill).

* DS_Store