Consistently frame the target element (#2799)
* Avoid copy-pasting the snapshot-testing.js file

We need it in each project we do snapshot testing in, but
I'm unaware of a way to load it from anywhere but the current
project's directory.

* Spotless

* Don't break configuration caching

* Consistently frame the target element

Support a target element that isn't already a child
of the document's root element.

When snapshotting, wrap the element in a frame, and
then remove the element from that frame afterwards.
This avoids mutating the passed-in element as a
side-effect of snapshotting.

* apiDump

* PR feedback

* apiDump

---------

Co-authored-by: Jesse Wilson <jwilson@squareup.com>