First pass at HTML Box (#2811)
* Use Helvetica or Robot at 17px for all snapshot tests

This brings metrics on iOS, Android and web as close
together as I can manage it.

Roboto in Chrome doesn't work as well as I would have
wanted without a download, so I'm using Helvetica there.

* Spotless

* Redo snapshots after rebase

* First pass at HTML Box

This wraps each child in an extra <div>, which makes a
bunch of the layout work easier. It also avoids hazards
where a parent and child compete over ownership of fields
like 'margin'.

* Update redwood-layout-dom/src/commonMain/kotlin/app/cash/redwood/layout/dom/HTMLElementBox.kt

Co-authored-by: Jake Wharton <jw@squareup.com>

* Update redwood-layout-dom/src/commonMain/kotlin/app/cash/redwood/layout/dom/HTMLSpacer.kt

Co-authored-by: Jake Wharton <jw@squareup.com>

* Update redwood-layout-dom/src/commonMain/kotlin/app/cash/redwood/layout/dom/HTMLFlexContainer.kt

Co-authored-by: Jake Wharton <jw@squareup.com>

---------

Co-authored-by: Jesse Wilson <jwilson@squareup.com>
Co-authored-by: Jake Wharton <jw@squareup.com>